refactor(index): split git history into cohesive siblings
Decompose git-history preparation, application, repository reads, queries, blame caching, and tests behind the existing compatibility surface.

Preserve rows-first cursor publication, append and rebuild selection, incomplete-history safeguards, repository scoping, and blame cache invalidation.

Fixes #921
